fix linter error Module has too many lines
split code into 2 modules
